Use non-root user in a Energi Core Node container
> Note: if `energi3-docker-compose` was used before, please read
> instructions at the end of this commit message before pulling these
> changes.

`Dockerfile` was updated to add user and group `nrgstaker`, set the
working directory to `/home/nrgstaker/energi3` and change owner for
added scripts to `nrgstaker` user and group.

Changes in `docker-compose.yml`:

- `secrets` was moved before `services`;
- `image` was added to `core` service;
- volume `core-data` was added;
- changed how data directory volume is specified for `core` service.

`scripts/bootstrap-chaindata.sh` was updated to use `~/` to move to the
home directory.

Script `setup.sh` and directory `setup` were added for first time setup
process.

Directory `volumes` was removed as it is not necessary anymore as now a
Docker volume is used for Energi data directory.

`README.md` was updated to reflect changes regarding first time setup.

`.gitignore` was added to exclude `docker-compose.override.yml` from the
repository.

---

If `energi3-docker-compose` was used before there are two options:

- pull changes and start everything from scratch by following
  instructions in `README.md`;
- a much faster way would be moving Energi data directory to a volume:
  - pull changes;
  - execute the following command from the root directory (it may take
    a while to finish):

    ``` sh
    docker-compose down && docker-compose run\
     --entrypoint '/bin/sh -c "\
      mv /root/.energicore3/* /home/nrgstaker/.energicore3/ &&
      chown -R nrgstaker:nrgstaker /home/nrgstaker/.energicore3"'\
     --rm\
     --user=root\
     --volume="${PWD}/volumes/root/.energicore3:/root/.energicore3"\
     core &&
    docker-compose up --detach
    ```

    this will stop and remove the Energi Core Node container, run a
    temporary container as root, mount `volumes/root/.energicore`, move
    its contents to the data directory volume, change ownership there
    and recreate and start the Core Node container.

    > `sudo` might be necessary before each `docker-compose`.
    >
    > It is allways advisable to make backups before data operations.