Fix PID retrieval in function `GET_NODE_INFO`
`nodemon/scripts/nodemon.sh` was updated to fix Energi Core Node Monitor
process PID retrieval in the function `GET_NODE_INFO`. Without detecting
the PID various types of messages were not sent.

A note on uptime value in informational messages. As originally
`nodemon.sh` is intended to run on the same server as Energi Core Node,
then uptime value is meant to be uptime of Core. In
`energi3-docker-compose` this value is the uptime of the cron process in
the Energi Core Node Monitor container as Core and Monitor run in separate
Docker containers.

This commit possibly resolves the issue #9.