Add single source of Energi Core Node version
Manual action is required for existing setups (see the last paragraph).

With these changes a single source of Energi Core Node version was
added: `.env` (and `.env.template.env`).

Helper script was updated to check if `.env` exists and to create it
with content of `.env.template.env` when necessary. `update` subcommand
was updated to add the possibility to specify a version.

The main `ReadMe.md` was updated to describe the new functionality,
reformat its structure and update Energi documentation links.

After pulling these changes the following actions has to be done
manually for existing setups:

- `docker-compose.override.yml` must be updated:

  - `build: nodemon` must be replaced with

    ``` yml
        build:
          args:
            ENERGI_VERSION: ${ENERGI_VERSION}
          context: nodemon
    ```

  - `image: energi-nodemon:3.1.0-1.1.1-0.0` must be replaced with
    `image: energi-nodemon:${ENERGI_VERSION}-1.1.1-0.0`.

- if `.env` is existing, `ENERGI_VERSION=v3.1.0` must be added to it;
  if `.env` does not exist, it must be created by copying
  `env.template.env` and changing value for `COMPOSE_PROJECT_NAME` to
  the current stack name (the name of the directory this repository was
  cloned into (by default `energi3-docker-compose`)).