Rename `.env.template.env` to `.env`
`.env.template.env` was renamed to `.env` as there should not be need to
use other values than the default. Even if there would be a need to
override the default values, it can be done in various ways.

The most possible scenario would be to use different Energi version than
one specified in `.env` as `ENERGI_VERSION` (for example, a new version
has been released but the repository has not been yet updated to reflect
that). It can be done in the following ways:

- use helper script (recommended):
  `./e3dc update [core|monitor] v3.1.1`;
- set environment variable in terminal: `export ENERGI_VERSION=v3.1.1`;

  setting environment variable before command *might* also work; for
  example:

  - `env ENERGI_VERSION=v3.1.1 ./e3dc update [core|monitor]`
  - `ENERGI_VERSION=v3.1.1 docker-compose up --build --detach
    [core|monitor]`;

- `.env` can be copied (for example, to `.env-override`) and modified,
  then it can be specified for `docker-compose` commands like this:
  `docker-compose --env-file .env-override up --build --detach`;
- `.env` can be modified directly, but changes will be overwritten at
  `git pull` or they will have to be manually reverted before
  `git pull` depending on the chosen pull/merge strategy.

Other `.env` values can be changed in the aforementioned ways (except
`./e3dc update [core|monitor] v3.1.1` as it changes value only for
`ENERGI_VERSION`).

It is possible that the existing `.env` should be removed before pulling
these changes.