Display NRG amount also in the configured currency
A new feature is introduced with this commit - a configurable option to
display NRG amount in messages also in the configured currency.

To use this feature, `nodemon/.env` has to be updated to add a new row:
`BALANCE_IN_CURRENCY=yes`. Then Energi Core Node Monitor container has
to be rebuilt and recreated: `./helper update monitor`.